Helper to clone tasks moved to PhysicsTools.PatAlgos.tools.helpers
